perf(ci): rebalance seven hosted lanes without reducing proof
